refactor: major code quality improvement and overhaul
